Move App from base to __init__.py
The base.py source file was used before refactoring it is much nicer to
have App in the main module instead.

The SimplelineError exception was moved to its own source file. This
change is required to fix cyclic imports.
