+++ examples/entry_counter/entry_counter.py
@@ -1,6 +1,6 @@
 #!/bin/python3
 
-from simpleline.base import App
+from simpleline import App
 from simpleline.render import InputState
 from simpleline.render.screen import UIScreen
 from simpleline.render.widgets import TextWidget, CenterWidget
