+++ examples/help/help.py
@@ -1,6 +1,6 @@
 #!/bin/python3
 
-from simpleline.base import App
+from simpleline import App
 from simpleline.render.adv_widgets import HelpScreen
