+++ examples/ignore_continue/ignore_continue.py
@@ -1,6 +1,6 @@
 #!/bin/python3
 
-from simpleline.base import App
+from simpleline import App
 from simpleline.render import InputState
 from simpleline.render.prompt import Prompt
 from simpleline.render.screen import UIScreen
