+++ simpleline/render/adv_widgets.py
@@ -19,7 +19,7 @@
 
 import sys
 
-from simpleline.base import App
+from simpleline import App
 from simpleline.render import widgets, InputState
 from simpleline.render.prompt import Prompt
 from simpleline.render.screen import UIScreen
