+++ tests/render_screen_test.py
@@ -22,7 +22,7 @@ import unittest
 from io import StringIO
 from unittest import mock
 
-from simpleline.base import App
+from simpleline import App
 from simpleline.render import InputState, RenderUnexpectedError
 from simpleline.render.screen import UIScreen
 from tests import schedule_screen_and_run, calculate_separator
